T34: refactor PrayerScreen.kt to MainScreen.kt
